requirements.txt
requests
tenacity
numpy
matplotlib
scikit-learn
chardet 
markdown 
beautifulsoup4 
tqdm 
PyPDF2 
python-docx
faiss-cpu
sentence-transformers
rank-bm25
fastapi
uvicorn
python-dotenv
pydantic
starlette
torch
pillow
pymupdf
pytesseract
git+https://github.com/openai/CLIP.git

docling
unstructured[pdf]
paddleocr
llama-index
pdfplumber
transformers
seaborn
raganything[all]